Refatoração de DTOs e ajustes em namespaces e serviços
- Adicionadas novas classes `AuthResponse` e `TeamResponse` em arquivos separados no namespace `CartolaLigas.DTOs.Response`.
- Movidas classes relacionadas a limites de competições para arquivos separados no namespace `CartolaLigas.Models.Cartola`.
- Alterados namespaces de `User` e `Time` para `CartolaLigas.Models`.
- Substituídas referências de `TeamDTO`, `TeamCartolaDTO` e `UserDTO` para refletir a nova estrutura de classes.
- Atualizada a lógica em componentes Razor e serviços para suportar as mudanças.
- Adicionada funcionalidade de edição de ligas e busca por ID em `LigasService`.
