+++ CartolaLigas.csproj
@@ -18,4 +18,8 @@
     <ServiceWorker Include="wwwroot\service-worker.js" PublishedContent="wwwroot\service-worker.published.js" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="DTOs\Request\" />
+  </ItemGroup>
+
 </Project>
