+++ DTOs/Response/AuthResponse.cs
@@ -1,0 +1,10 @@
+﻿using CartolaLigas.Models;
+
+namespace CartolaLigas.DTOs.Response
+{
+    public class AuthResponse
+    {
+        public string token { get; set; }
+        public User record { get; set; }
+    }
+}
