+++ DTOs/Response/MercadoResponse.cs
@@ -1,73 +1,5 @@
 using System.Text.Json.Serialization;
-
-public class Fechamento
-{
-    [JsonPropertyName("dia")]
-    public int Dia { get; set; }
-
-    [JsonPropertyName("mes")]
-    public int Mes { get; set; }
-
-    [JsonPropertyName("ano")]
-    public int Ano { get; set; }
-
-    [JsonPropertyName("hora")]
-    public int Hora { get; set; }
-
-    [JsonPropertyName("minuto")]
-    public int Minuto { get; set; }
-
-    [JsonPropertyName("timestamp")]
-    public long Timestamp { get; set; }
-}
-
-public class LimitesCompeticao
-{
-    [JsonPropertyName("total_confronto_pro")]
-    public int TotalConfrontoPro { get; set; }
-
-    [JsonPropertyName("total_confronto_free")]
-    public int TotalConfrontoFree { get; set; }
-
-    [JsonPropertyName("criacao_confronto_pro")]
-    public int CriacaoConfrontoPro { get; set; }
-
-    [JsonPropertyName("criacao_confronto_free")]
-    public int CriacaoConfrontoFree { get; set; }
-}
-
-public class Free
-{
-    [JsonPropertyName("criacao")]
-    public int Criacao { get; set; }
-
-    [JsonPropertyName("participacao")]
-    public int Participacao { get; set; }
-}
-
-public class Pro
-{
-    [JsonPropertyName("criacao")]
-    public int Criacao { get; set; }
-
-    [JsonPropertyName("participacao")]
-    public int Participacao { get; set; }
-}
-
-public class PontosCorridos
-{
-    [JsonPropertyName("free")]
-    public Free Free { get; set; }
-
-    [JsonPropertyName("pro")]
-    public Pro Pro { get; set; }
-}
-
-public class LimitesCompeticoes
-{
-    [JsonPropertyName("pontos_corridos")]
-    public PontosCorridos PontosCorridos { get; set; }
-}
+using CartolaLigas.Models.Cartola;
 
 public class MercadoResponse
 {
