+++ DTOs/Response/TeamResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace CartolaLigas.DTOs.Response
+{
+    public class TeamResponse
+    {
+    }
+}
