+++ Models/Cartola/PontosCorridos.cs
@@ -1,0 +1,13 @@
+﻿using System.Text.Json.Serialization;
+
+namespace CartolaLigas.Models.Cartola
+{
+    public class PontosCorridos
+    {
+        [JsonPropertyName("free")]
+        public Free Free { get; set; }
+
+        [JsonPropertyName("pro")]
+        public Pro Pro { get; set; }
+    }
+}
