+++ Models/Cartola/Time.cs
@@ -1,8 +1,8 @@
 ﻿using System.Text.Json.Serialization;
 
-namespace CartolaLigas.DTOs
+namespace CartolaLigas.Models.Cartola
 {
-    public class TeamCartolaDTO
+    public class Time
     {
         [JsonPropertyName("url_escudo_svg")]
         public string UrlEscudoSvg { get; set; }
@@ -38,25 +38,5 @@ namespace CartolaLigas.DTOs
         public double PontosCampeonato { get; set; }
         [JsonPropertyName("patrimonio")]
         public double Patrimonio { get; set; }
-    }
-
-    public class TeamDTO
-    {
-        public string CollectionId { get; set; }
-        public string CollectionName { get; set; }
-        public string Id { get; set; }
-        public string Slug { get; set; }
-        public string Name { get; set; }
-        [JsonPropertyName("user_id")]
-        public string UserId { get; set; }
-        [JsonPropertyName("cartola_time_id")]
-        public long CartolaTimeId { get; set; }
-        [JsonPropertyName("nome_cartola")]
-        public string NomeCartola { get; set; }
-        public string Created { get; set; }
-        public string Updated { get; set; }
-    }
-    public class TeamResponse
-    {
     }
 }
