+++ Pages/Teams/Teams.razor
@@ -1,6 +1,7 @@
 ﻿@page "/teams"
 @using System.Text.Json.Serialization
 @using CartolaLigas.DTOs
+@using CartolaLigas.Models.Cartola
 @inject TimeService TimeService
 @inject MercadoService MercadoService
 
@@ -54,7 +55,7 @@
 
 @code {
 	private string searchQuery;
-	private List<TeamCartolaDTO> teams;
+	private List<Time> teams;
 	private bool isSearching = false;
 	private string? errorMessage = null;
