+++ Services/TimeService.cs
@@ -1,6 +1,6 @@
 ﻿using System.IdentityModel.Tokens.Jwt;
 using System.Net.Http.Json;
-using CartolaLigas.DTOs;
+using CartolaLigas.Models.Cartola;
 using CartolaLigas.Providers;
 using Microsoft.JSInterop;
 
@@ -11,7 +11,7 @@ namespace CartolaLigas.Services
         private readonly HttpClient _httpClient;
         private readonly IJSRuntime _jSRuntime;
         private readonly MercadoService _mercadoService;
-        private TeamDTO? _cachedTeam; // Cache em memória para o time
+        private Models.Time? _cachedTeam; // Cache em memória para o time
 
 
         public TimeService(HttpClient httpClient, IJSRuntime jSRuntime, MercadoService mercadoService)
@@ -20,9 +20,9 @@ namespace CartolaLigas.Services
             _jSRuntime = jSRuntime;
             _mercadoService = mercadoService;
         }
-        public async Task<List<TeamCartolaDTO>> SearchTeams(string query, bool details = false)
+        public async Task<List<Models.Cartola.Time>> SearchTeams(string query, bool details = false)
         {
-            var teams = await _httpClient.GetFromJsonAsync<List<TeamCartolaDTO>>($"https://api.ligas.ehtudo.app/webhook/cartola/v1/busca?query={query}");
+            var teams = await _httpClient.GetFromJsonAsync<List<Models.Cartola.Time>>($"https://api.ligas.ehtudo.app/webhook/cartola/v1/busca?query={query}");
             if (details)
             {
                 if (teams != null && teams[0].Nome != null)
@@ -34,7 +34,7 @@ namespace CartolaLigas.Services
             return teams;
         }
 
-        private async Task EnrichTeamsWithDetails(List<TeamCartolaDTO> teams)
+        private async Task EnrichTeamsWithDetails(List<Models.Cartola.Time> teams)
         {
             var mercadoStatus = await _mercadoService.GetMercadoStatus();
             if (!_mercadoService.IsEmManutencao())
@@ -43,7 +43,7 @@ namespace CartolaLigas.Services
                 {
                     try
                     {
-                        var teamDetails = await _httpClient.GetFromJsonAsync<TeamCartolaDTO>($"https://bypass.ehtudo.app/https://api.cartola.globo.com/time/id/{team.TimeId}");
+                        var teamDetails = await _httpClient.GetFromJsonAsync<Models.Cartola.Time>($"https://bypass.ehtudo.app/https://api.cartola.globo.com/time/id/{team.TimeId}");
                         if (teamDetails != null)
                         {
                             team.PontosCampeonato = teamDetails.PontosCampeonato;
@@ -60,10 +60,10 @@ namespace CartolaLigas.Services
             }
         }
 
-        public async Task<TeamDTO> AddTime(TeamDTO teamDTO)
+        public async Task<Models.Time> AddTime(Models.Time time)
         {
             // Verificar se o time já existe na base de dados
-            var existingTeamResponse = await _httpClient.GetFromJsonAsync<TimeResponse>($"https://api.ligas.ehtudo.app/api/collections/times/records?filter=(cartola_time_id={teamDTO.CartolaTimeId})");
+            var existingTeamResponse = await _httpClient.GetFromJsonAsync<TimeResponse>($"https://api.ligas.ehtudo.app/api/collections/times/records?filter=(cartola_time_id={time.CartolaTimeId})");
             if (existingTeamResponse != null && existingTeamResponse.items != null && existingTeamResponse.items.Count > 0)
             {
                 // Retornar o time existente
@@ -71,10 +71,10 @@ namespace CartolaLigas.Services
             }
 
             _httpClient.DefaultRequestHeaders.Clear();
-            var response = await _httpClient.PostAsJsonAsync("https://api.ligas.ehtudo.app/api/collections/times/records", teamDTO);
+            var response = await _httpClient.PostAsJsonAsync("https://api.ligas.ehtudo.app/api/collections/times/records", time);
             if (response.IsSuccessStatusCode)
             {
-                var team = await response.Content.ReadFromJsonAsync<TeamDTO>();
+                var team = await response.Content.ReadFromJsonAsync<Models.Time>();
                 List<Object> timeRodadaQueue = new List<Object>();
                 await _mercadoService.GetMercadoStatus();
                 for (int i = 1; i < _mercadoService.GetRodadaAtual(); i++)
@@ -86,11 +86,11 @@ namespace CartolaLigas.Services
             }
             return null;
         }
-        public async Task<TeamDTO> AddOwnTeam(TeamDTO teamDTO, string? token = null)
+        public async Task<Models.Time> AddOwnTeam(Models.Time time, string? token = null)
         {
             //obter o authToken do localStorage
             var authToken = token ?? await _jSRuntime.InvokeAsync<string>("localStorage.getItem", "authToken");
-            var addedTeam = await AddTime(teamDTO);
+            var addedTeam = await AddTime(time);
             if (authToken != null || addedTeam != null)
             {
                 _httpClient.DefaultRequestHeaders.Clear();
@@ -100,13 +100,13 @@ namespace CartolaLigas.Services
                 var response = await _httpClient.PostAsJsonAsync("https://api.ligas.ehtudo.app/webhook/ligas/v1/time/addOwnTime", addedTeam);
                 if (response.IsSuccessStatusCode)
                 {
-                    var team = await response.Content.ReadFromJsonAsync<TeamDTO>();
+                    var team = await response.Content.ReadFromJsonAsync<Models.Time>();
                     return team;
                 }
             }
             return null;
         }
-        public async Task RemoveOwnTeam(TeamDTO teamDTO, string? token = null)
+        public async Task RemoveOwnTeam(Models.Time time, string? token = null)
         {
             //obter o authToken do localStorage
             var authToken = token ?? await _jSRuntime.InvokeAsync<string>("localStorage.getItem", "authToken");
@@ -114,10 +114,10 @@ namespace CartolaLigas.Services
             {
                 _httpClient.DefaultRequestHeaders.Clear();
                 _httpClient.DefaultRequestHeaders.Add("Authorization", authToken);
-                var response = await _httpClient.PostAsJsonAsync($"https://api.ligas.ehtudo.app/webhook/ligas/v1/time/removeOwnTime",teamDTO);
+                var response = await _httpClient.PostAsJsonAsync($"https://api.ligas.ehtudo.app/webhook/ligas/v1/time/removeOwnTime",time);
                 if (response.IsSuccessStatusCode)
                 {
-                    var team = await response.Content.ReadFromJsonAsync<TeamDTO>();
+                    var team = await response.Content.ReadFromJsonAsync<Models.Time>();
                     ClearCache();
                 }
                 else
@@ -128,7 +128,7 @@ namespace CartolaLigas.Services
             }
         }
 
-        public async Task<TeamDTO> Time(string? token = null)
+        public async Task<Models.Time> Time(string? token = null)
         {
             // Verificar se o time já está no cache
             if (_cachedTeam != null)
@@ -169,7 +169,6 @@ namespace CartolaLigas.Services
         private async Task adicionarPontuacaoTime()
         {
             var response = await _httpClient.GetFromJsonAsync<Object>($"https://api.ligas.ehtudo.app/webhook/ligas/v1/time");
-
         }
 
         public class ErrorResponse
@@ -192,7 +191,7 @@ namespace CartolaLigas.Services
 
         public class TimeResponse
         {
-            public List<TeamDTO>? items { get; set; }
+            public List<Models.Time>? items { get; set; }
             public int page { get; set; }
             public int perPage { get; set; }
             public int totalItems { get; set; }
